requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Kafka
kafka-python==2.0.2

# Database
pymongo==4.6.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-json-logger==2.0.7

